refactor(cli): split doctor into DI doctor-core to fix Linux coverage gate
`test:coverage:doctor` failed deterministically on Linux CI (`bun test
... --coverage-threshold-lines=80` exit 2) after the PR-445 expansion of
`doctor.test.ts`. Symptoms:

  All files                       |   78.09 |   75.51 |
   src/commands/doctor.ts         |   78.95 |   82.72 | ...
   src/paths.ts                   |  100.00 |   45.07 | 44,52-56,67-98
   ...                            |   ...   |   ...   |

The real target (`doctor.ts`) was fine at ~83 %, but `paths.ts` has three
platform branches (win32 / darwin / linux) and naturally line-covers
~45 % on any single OS. PR 445 added `runDoctor` tests that route through
`cli-mocks.ts`; even with `mock.module` in place, Bun's coverage
instrumenter still walks every transitively imported source, so
`paths.ts` (and the test/helpers) dragged `All files` below the 80 %
per-file threshold.

Fix mirrors the `repl-core.ts` <-> `repl.ts` split landed in the previous
commit (same shape as `gw-state-helpers.ts` <-> `gateway-process.ts`):

- `doctor-core.ts` (new) — every export function (`worstHealthSeverity`,
  `healthStateMark`, `doctorPrintConfigValidation`,
  `doctorPrintIndexFromSnapshot`, `doctorPrintHealthFromSnapshot`,
  `doctorVoiceLines`, and `runDoctor`). `runDoctor` now takes a
  `DoctorCoreDeps` parameter (getCliPlatformPaths / readGatewayState /
  isProcessAlive / gatewayStatePath / makeClient). Imports NONE of the
  production-only modules as values; signatures mirror
  `gateway-process.ts` exactly so the shim wires `productionDeps`
  without conversion.
- `doctor.ts` — thin production-wiring shim. Statically imports
  `paths.ts`, `gateway-process.ts`, `ipc-client/index.ts`, builds
  `productionDeps`, delegates to `coreRunDoctor(args, productionDeps)`.
  Coverage-exempt (no testable branches).
- `doctor-core.test.ts` (new) — moves the entire suite (47 tests) over.
  Does NOT import `cli-mocks.ts`; pure helper tests + `runDoctor`
  dispatcher tests both use injected `DoctorCoreDeps` fakes.
  `afterEach` now explicitly zeroes `process.exitCode` before
  restoring it — leaving Bun's process exit code at the value set by
  one of the dispatcher tests (~2) makes the whole gate exit non-zero
  even when every test passes; explicit reset keeps the runner clean.
- `doctor-voice.test.ts` — updated to import `DoctorEnv`,
  `DoctorVoiceConfig`, `doctorVoiceLines` from `doctor-core.ts` (the
  shim no longer re-exports them).
- `doctor.test.ts` — deleted.
- Coverage exclusions: `doctor.ts` added to
  `scripts/coverage-floor/exclusions.ts` and `sonar-project.properties`
  (audit:exclusion-parity green).

Linux verification (5x via `docker run oven/bun:latest`):
  `bun run test:coverage:doctor` -> EXIT 0, all 47 tests pass, `All files`
  >= 83 %, `doctor-core.ts` at 82.5 %.
