+++ packages/cli/src/commands/doctor-core.ts
@@ -1,0 +1,385 @@
+// packages/cli/src/commands/doctor-core.ts
+//
+// Dependency-injected `nimbus doctor` logic. This file is the testable twin of
+// the thin `doctor.ts` wrapper (same split as `repl-core.ts` ↔ `repl.ts`,
+// `gw-state-helpers.ts` ↔ `gateway-process.ts`).
+//
+// Why the split exists: the colocated test must run alongside the shared CLI
+// mock harness (`cli-mocks.ts`), which installs process-global `mock.module`
+// registrations for `ipc-client/index.ts`, `lib/gateway-process.ts`, and
+// `@clack/prompts`. Even with the mocks in place Bun's coverage instrumenter
+// still walks every transitively-imported source — including `paths.ts`,
+// whose three platform branches naturally line-cover ~45 % on any single OS.
+// On the Linux CI gate that drags `All files` below the 80 % threshold and
+// fails `bun run test:coverage:doctor`, even though the real target
+// (`doctor.ts`) sits at ~83 %.
+//
+// `doctor-core.ts` therefore imports NONE of the production-only modules as
+// values — only Node builtins and type-only references — and takes its real
+// dependencies via a `DoctorCoreDeps` object. The test imports this file
+// directly and injects fakes, so it never needs `cli-mocks.ts`; the
+// module-under-test has a clean graph and a stable coverage scope on every
+// platform. `doctor.ts` supplies the production deps and is the
+// (coverage-exempt) wiring shim.
+import { existsSync, readFileSync } from "node:fs";
+import { platform } from "node:os";
+import { join } from "node:path";
+
+import type { IPCClient } from "../ipc-client/index.ts";
+import type { CliPlatformPaths } from "../paths.ts";
+
+const LINUX_SECRET_TOOL_HINT =
+  "secret-tool not found. Install libsecret-tools (Debian/Ubuntu) or libsecret (Fedora/Arch) to use the OS vault on Linux.";
+
+const MIN_BUN_MAJOR = 1;
+const MIN_BUN_MINOR = 2;
+
+function bunVersionOk(): boolean {
+  const m = /^(\d+)\.(\d+)\./.exec(Bun.version);
+  if (m === null) {
+    return true;
+  }
+  const major = Number(m[1]);
+  const minor = Number(m[2]);
+  return major > MIN_BUN_MAJOR || (major === MIN_BUN_MAJOR && minor >= MIN_BUN_MINOR);
+}
+
+type ConnectorHealthRow = {
+  connectorId?: unknown;
+  state?: unknown;
+};
+
+/** Minimal shape of the recorded gateway state `runDoctor` reads. */
+export interface DoctorGatewayState {
+  readonly socketPath: string;
+  readonly pid: number;
+}
+
+/**
+ * The real dependencies `doctor.ts` wires in. Injected (never statically
+ * imported here) so this module stays out of the cli-mocks blast radius.
+ * Signatures mirror `packages/cli/src/lib/gateway-process.ts` exactly so the
+ * production shim can wire `productionDeps` without conversion shims.
+ */
+export interface DoctorCoreDeps {
+  readonly getCliPlatformPaths: () => CliPlatformPaths;
+  readonly readGatewayState: (paths: CliPlatformPaths) => Promise<DoctorGatewayState | undefined>;
+  readonly isProcessAlive: (pid: number) => boolean;
+  readonly gatewayStatePath: (paths: CliPlatformPaths) => string;
+  readonly makeClient: (socketPath: string) => IPCClient;
+}
+
+/**
+ * Test entry point — invoked by helpers in `doctor-core.ts` and the colocated
+ * `doctor-core.test.ts`. Do not call from other command files.
+ */
+export function worstHealthSeverity(rows: ConnectorHealthRow[]): "ok" | "warn" | "fail" {
+  let worst: "ok" | "warn" | "fail" = "ok";
+  for (const r of rows) {
+    const st = typeof r.state === "string" ? r.state : "";
+    if (st === "unauthenticated" || st === "error") {
+      worst = "fail";
+    } else if (st === "degraded" || st === "rate_limited") {
+      if (worst === "ok") {
+        worst = "warn";
+      }
+    }
+  }
+  return worst;
+}
+
+/**
+ * Test entry point — invoked by helpers in `doctor-core.ts` and the colocated
+ * `doctor-core.test.ts`. Do not call from other command files.
+ */
+export function healthStateMark(st: string): string {
+  if (st === "healthy" || st === "paused") {
+    return "[ok]";
+  }
+  if (st === "unauthenticated" || st === "error") {
+    return "[fail]";
+  }
+  return "[warn]";
+}
+
+function doctorPrintBunCheck(): number {
+  console.log(`Runtime: Bun ${Bun.version}`);
+  if (bunVersionOk()) {
+    console.log("[ok] Bun version meets minimum.");
+    return 0;
+  }
+  console.log(
+    `[fail] Nimbus expects Bun >= ${String(MIN_BUN_MAJOR)}.${String(MIN_BUN_MINOR)} (see repository README).`,
+  );
+  return 2;
+}
+
+function doctorPrintVaultCheck(): number {
+  if (platform() === "linux") {
+    if (Bun.which("secret-tool") === null) {
+      console.log(`[fail] Vault: ${LINUX_SECRET_TOOL_HINT}`);
+      return 2;
+    }
+    console.log("[ok] Vault: secret-tool is on PATH.");
+    return 0;
+  }
+  console.log(`[ok] Vault: OS-native store (${platform()}) — no Linux secret-tool check.`);
+  return 0;
+}
+
+/**
+ * Test entry point — invoked by `doctorRunGatewayRpcs` and the colocated
+ * `doctor-core.test.ts`. Do not call from other command files.
+ */
+export function doctorPrintConfigValidation(val: {
+  ok: boolean;
+  errors: string[];
+  warnings: string[];
+}): number {
+  let exit = 0;
+  if (val.warnings.length > 0) {
+    for (const w of val.warnings) {
+      console.log(`[warn] Config: ${w}`);
+    }
+    exit = 1;
+  }
+  if (!val.ok) {
+    for (const e of val.errors) {
+      console.log(`[fail] Config: ${e}`);
+    }
+    return 2;
+  }
+  if (val.errors.length === 0 && val.warnings.length === 0) {
+    console.log("[ok] Config: valid.");
+  }
+  return exit;
+}
+
+/**
+ * Test entry point — invoked by `doctorRunGatewayRpcs` and the colocated
+ * `doctor-core.test.ts`. Do not call from other command files.
+ */
+export function doctorPrintIndexFromSnapshot(snap: { index?: { totalItems?: unknown } }): number {
+  const total = snap.index?.totalItems;
+  const nItems =
+    typeof total === "number" && Number.isFinite(total) ? Math.max(0, Math.floor(total)) : 0;
+  if (nItems === 0) {
+    console.log("[warn] Index: zero items — run connector sync after auth.");
+    return 1;
+  }
+  console.log(`[ok] Index: ${String(nItems)} items.`);
+  return 0;
+}
+
+/**
+ * Test entry point — invoked by `doctorRunGatewayRpcs` and the colocated
+ * `doctor-core.test.ts`. Do not call from other command files.
+ */
+export function doctorPrintHealthFromSnapshot(snap: { connectorHealth?: unknown }): number {
+  const healthRaw = snap.connectorHealth;
+  const health: ConnectorHealthRow[] = Array.isArray(healthRaw)
+    ? (healthRaw as ConnectorHealthRow[])
+    : [];
+  if (health.length === 0) {
+    console.log("[warn] Connectors: none registered.");
+    return 1;
+  }
+  console.log("Connector health:");
+  for (const h of health) {
+    const id = typeof h.connectorId === "string" ? h.connectorId : "?";
+    const st = typeof h.state === "string" ? h.state : "?";
+    const mark = healthStateMark(st);
+    console.log(`  ${mark} ${id}: ${st}`);
+  }
+  const sev = worstHealthSeverity(health);
+  return sev === "ok" ? 0 : 1;
+}
+
+async function doctorRunGatewayRpcs(client: IPCClient): Promise<number> {
+  const ping = await client.call<{ uptime?: number }>("gateway.ping", {});
+  const uptime = typeof ping.uptime === "number" && Number.isFinite(ping.uptime) ? ping.uptime : 0;
+  console.log(`[ok] Gateway: IPC OK (uptime ~${String(Math.round(uptime / 1000))}s).`);
+
+  const val = await client.call<{ ok: boolean; errors: string[]; warnings: string[] }>(
+    "config.validate",
+    {},
+  );
+  let exit = doctorPrintConfigValidation(val);
+
+  const snap = await client.call<{
+    index?: { totalItems?: unknown };
+    connectorHealth?: unknown;
+  }>("diag.snapshot", {});
+  exit = Math.max(exit, doctorPrintIndexFromSnapshot(snap));
+  exit = Math.max(exit, doctorPrintHealthFromSnapshot(snap));
+  return exit;
+}
+
+/**
+ * The full `nimbus doctor` dispatcher, dependency-injected. `doctor.ts`
+ * builds the production deps and delegates here; tests pass a `DoctorCoreDeps`
+ * fake so the run never touches the real IPC, vault, or gateway state files.
+ */
+export async function runDoctor(_args: string[], deps: DoctorCoreDeps): Promise<void> {
+  const paths = deps.getCliPlatformPaths();
+  let exit = 0;
+  exit = Math.max(exit, doctorPrintBunCheck());
+
+  console.log(`Data dir: ${paths.dataDir}`);
+  console.log(`Gateway state file: ${deps.gatewayStatePath(paths)}`);
+
+  exit = Math.max(exit, doctorPrintVaultCheck());
+
+  const voiceCfg = loadVoiceConfigFromDir(paths.configDir);
+  const voiceLines = doctorVoiceLines(voiceCfg, {
+    which: (n) => Bun.which(n),
+    platform: platform() as "win32" | "darwin" | "linux",
+  });
+  for (const line of voiceLines) {
+    console.log(line);
+    if (line.startsWith("[fail]")) exit = Math.max(exit, 2);
+    else if (line.startsWith("[warn]")) exit = Math.max(exit, 1);
+  }
+
+  const state = await deps.readGatewayState(paths);
+  if (state !== undefined && deps.isProcessAlive(state.pid)) {
+    const client = deps.makeClient(state.socketPath);
+    try {
+      await client.connect();
+      exit = Math.max(exit, await doctorRunGatewayRpcs(client));
+    } catch (e) {
+      const msg = e instanceof Error ? e.message : String(e);
+      console.log(`[fail] Gateway: IPC failed — ${msg}`);
+      exit = Math.max(exit, 2);
+    } finally {
+      await client.disconnect().catch(() => {});
+    }
+  } else if (state === undefined) {
+    console.log("[fail] Gateway: not running (no gateway.json — start with: nimbus start).");
+    exit = Math.max(exit, 2);
+  } else {
+    console.log(
+      `[fail] Gateway: stale state (pid ${String(state.pid)} is not running) — try nimbus stop or remove the state file.`,
+    );
+    exit = Math.max(exit, 2);
+  }
+
+  process.exitCode = exit;
+}
+
+// ─── Voice doctor helpers ────────────────────────────────────────────────────
+
+export type DoctorVoiceConfig = {
+  enabled: boolean;
+  whisperPath: string;
+  piperPath: string;
+  piperModel: string;
+};
+
+export type DoctorEnv = {
+  which: (name: string) => string | null;
+  platform: "win32" | "darwin" | "linux";
+};
+
+function doctorPiperLines(cfg: DoctorVoiceConfig, env: DoctorEnv): string[] {
+  if (cfg.piperPath === "" && cfg.piperModel === "") return [];
+  const lines: string[] = [];
+  const hasAbsPath = cfg.piperPath.includes("/") || cfg.piperPath.includes("\\");
+  const piperBinOk = cfg.piperPath !== "" && (hasAbsPath || env.which(cfg.piperPath) !== null);
+  if (!piperBinOk) {
+    lines.push(`[warn] Voice: piper_path is set but the binary was not found: ${cfg.piperPath}`);
+  }
+  if (cfg.piperModel === "") {
+    lines.push(
+      "[warn] Voice: piper_path is set but piper_model is empty — Piper TTS will not run.",
+    );
+  }
+  return lines;
+}
+
+export function doctorVoiceLines(cfg: DoctorVoiceConfig, env: DoctorEnv): string[] {
+  if (!cfg.enabled) return [];
+  const lines: string[] = [];
+
+  const whisperOk =
+    cfg.whisperPath !== "" || env.which("whisper-cli") !== null || env.which("main") !== null;
+  lines.push(
+    whisperOk
+      ? "[ok] Voice: whisper-cli is available."
+      : "[warn] Voice: whisper-cli not found on PATH and voice.whisper_path is unset — STT will not work.",
+  );
+
+  const ffmpegOk = env.which("ffmpeg") !== null;
+  lines.push(
+    ffmpegOk
+      ? "[ok] Voice: ffmpeg is on PATH."
+      : "[warn] Voice: ffmpeg not found on PATH — wake word detection requires ffmpeg for audio capture.",
+  );
+
+  if (env.platform === "darwin") {
+    lines.push("[ok] Voice: macOS `say` is always available.");
+  } else if (env.platform === "win32") {
+    lines.push("[ok] Voice: Windows SAPI via PowerShell is always available.");
+  } else {
+    const espeakOk = env.which("espeak-ng") !== null;
+    const spdSayOk = env.which("spd-say") !== null;
+    if (espeakOk) {
+      lines.push("[ok] Voice: Linux TTS via espeak-ng.");
+    } else if (spdSayOk) {
+      lines.push("[ok] Voice: Linux TTS via spd-say (espeak-ng preferred).");
+    } else {
+      lines.push(
+        "[warn] Voice: neither espeak-ng nor spd-say found on PATH — install one to enable TTS on Linux.",
+      );
+    }
+  }
+
+  lines.push(...doctorPiperLines(cfg, env));
+  return lines;
+}
+
+function applyVoiceKey(out: Partial<DoctorVoiceConfig>, key: string, val: string): void {
+  if (key === "enabled") out.enabled = val === "true";
+  else if (key === "whisper_path") out.whisperPath = val;
+  else if (key === "piper_path") out.piperPath = val;
+  else if (key === "piper_model") out.piperModel = val;
+}
+
+function parseVoiceTomlLines(lines: string[]): Partial<DoctorVoiceConfig> {
+  let inVoice = false;
+  const out: Partial<DoctorVoiceConfig> = {};
+  for (const line of lines) {
+    const hashIdx = line.indexOf("#");
+    const trimmed = (hashIdx >= 0 ? line.slice(0, hashIdx) : line).trim();
+    if (trimmed === "") continue;
+    if (trimmed.startsWith("[") && trimmed.endsWith("]")) {
+      inVoice = trimmed === "[voice]";
+      continue;
+    }
+    if (!inVoice) continue;
+    const eq = trimmed.indexOf("=");
+    if (eq <= 0) continue;
+    const key = trimmed.slice(0, eq).trim();
+    const valRaw = trimmed.slice(eq + 1).trim();
+    const val = valRaw.startsWith('"') && valRaw.endsWith('"') ? valRaw.slice(1, -1) : valRaw;
+    applyVoiceKey(out, key, val);
+  }
+  return out;
+}
+
+function loadVoiceConfigFromDir(configDir: string): DoctorVoiceConfig {
+  const tomlPath = join(configDir, "nimbus.toml");
+  const defaults: DoctorVoiceConfig = {
+    enabled: false,
+    whisperPath: "",
+    piperPath: "",
+    piperModel: "",
+  };
+  if (!existsSync(tomlPath)) return defaults;
+  try {
+    const src = readFileSync(tomlPath, "utf8");
+    return { ...defaults, ...parseVoiceTomlLines(src.split(/\r?\n/)) };
+  } catch {
+    return defaults;
+  }
+}
