+++ packages/cli/src/commands/doctor.ts
@@ -1,338 +1,34 @@
-import { existsSync, readFileSync } from "node:fs";
-import { platform } from "node:os";
-import { join } from "node:path";
-
+// packages/cli/src/commands/doctor.ts
+//
+// Thin production-wiring shim over `doctor-core.ts`. This file STATICALLY
+// imports the cli-mocks-mocked modules (`ipc-client`, `gateway-process`, and —
+// transitively via `paths.ts` — the platform-specific path resolver) to build
+// the real `DoctorCoreDeps`, which places it in Bun's mock-resolution +
+// coverage-instrumentation blast radius. It is therefore deliberately kept
+// logic-free and is coverage-exempt (scripts/coverage-floor/exclusions.ts +
+// sonar-project.properties), mirroring the `repl.ts` / `gateway-process.ts`
+// splits. All testable logic lives in `doctor-core.ts`, which the colocated
+// test exercises directly with injected fakes (no cli-mocks, stable coverage
+// scope on every platform).
 import { IPCClient } from "../ipc-client/index.ts";
 import { gatewayStatePath, isProcessAlive, readGatewayState } from "../lib/gateway-process.ts";
 import { getCliPlatformPaths } from "../paths.ts";
 
-const LINUX_SECRET_TOOL_HINT =
-  "secret-tool not found. Install libsecret-tools (Debian/Ubuntu) or libsecret (Fedora/Arch) to use the OS vault on Linux.";
+import { runDoctor as coreRunDoctor, type DoctorCoreDeps } from "./doctor-core.ts";
 
-const MIN_BUN_MAJOR = 1;
-const MIN_BUN_MINOR = 2;
-
-function bunVersionOk(): boolean {
-  const m = /^(\d+)\.(\d+)\./.exec(Bun.version);
-  if (m === null) {
-    return true;
-  }
-  const major = Number(m[1]);
-  const minor = Number(m[2]);
-  return major > MIN_BUN_MAJOR || (major === MIN_BUN_MAJOR && minor >= MIN_BUN_MINOR);
-}
-
-type ConnectorHealthRow = {
-  connectorId?: unknown;
-  state?: unknown;
+const productionDeps: DoctorCoreDeps = {
+  getCliPlatformPaths,
+  readGatewayState,
+  isProcessAlive,
+  gatewayStatePath,
+  makeClient: (socketPath) => new IPCClient(socketPath),
 };
 
 /**
- * Test entry point — invoked by helpers in `doctor.ts` and the colocated
- * `doctor.test.ts`. Do not call from other command files.
+ * Dispatcher entry point (`nimbus doctor`). Wires the production dependencies
+ * and delegates to the tested `doctor-core.ts` logic. Do not add logic here —
+ * add it to `doctor-core.ts` so it stays covered.
  */
-export function worstHealthSeverity(rows: ConnectorHealthRow[]): "ok" | "warn" | "fail" {
-  let worst: "ok" | "warn" | "fail" = "ok";
-  for (const r of rows) {
-    const st = typeof r.state === "string" ? r.state : "";
-    if (st === "unauthenticated" || st === "error") {
-      worst = "fail";
-    } else if (st === "degraded" || st === "rate_limited") {
-      if (worst === "ok") {
-        worst = "warn";
-      }
-    }
-  }
-  return worst;
-}
-
-/**
- * Test entry point — invoked by helpers in `doctor.ts` and the colocated
- * `doctor.test.ts`. Do not call from other command files.
- */
-export function healthStateMark(st: string): string {
-  if (st === "healthy" || st === "paused") {
-    return "[ok]";
-  }
-  if (st === "unauthenticated" || st === "error") {
-    return "[fail]";
-  }
-  return "[warn]";
-}
-
-function doctorPrintBunCheck(): number {
-  console.log(`Runtime: Bun ${Bun.version}`);
-  if (bunVersionOk()) {
-    console.log("[ok] Bun version meets minimum.");
-    return 0;
-  }
-  console.log(
-    `[fail] Nimbus expects Bun >= ${String(MIN_BUN_MAJOR)}.${String(MIN_BUN_MINOR)} (see repository README).`,
-  );
-  return 2;
-}
-
-function doctorPrintVaultCheck(): number {
-  if (platform() === "linux") {
-    if (Bun.which("secret-tool") === null) {
-      console.log(`[fail] Vault: ${LINUX_SECRET_TOOL_HINT}`);
-      return 2;
-    }
-    console.log("[ok] Vault: secret-tool is on PATH.");
-    return 0;
-  }
-  console.log(`[ok] Vault: OS-native store (${platform()}) — no Linux secret-tool check.`);
-  return 0;
-}
-
-/**
- * Test entry point — invoked by `doctorRunGatewayRpcs` and the colocated
- * `doctor.test.ts`. Do not call from other command files.
- */
-export function doctorPrintConfigValidation(val: {
-  ok: boolean;
-  errors: string[];
-  warnings: string[];
-}): number {
-  let exit = 0;
-  if (val.warnings.length > 0) {
-    for (const w of val.warnings) {
-      console.log(`[warn] Config: ${w}`);
-    }
-    exit = 1;
-  }
-  if (!val.ok) {
-    for (const e of val.errors) {
-      console.log(`[fail] Config: ${e}`);
-    }
-    return 2;
-  }
-  if (val.errors.length === 0 && val.warnings.length === 0) {
-    console.log("[ok] Config: valid.");
-  }
-  return exit;
-}
-
-/**
- * Test entry point — invoked by `doctorRunGatewayRpcs` and the colocated
- * `doctor.test.ts`. Do not call from other command files.
- */
-export function doctorPrintIndexFromSnapshot(snap: { index?: { totalItems?: unknown } }): number {
-  const total = snap.index?.totalItems;
-  const nItems =
-    typeof total === "number" && Number.isFinite(total) ? Math.max(0, Math.floor(total)) : 0;
-  if (nItems === 0) {
-    console.log("[warn] Index: zero items — run connector sync after auth.");
-    return 1;
-  }
-  console.log(`[ok] Index: ${String(nItems)} items.`);
-  return 0;
-}
-
-/**
- * Test entry point — invoked by `doctorRunGatewayRpcs` and the colocated
- * `doctor.test.ts`. Do not call from other command files.
- */
-export function doctorPrintHealthFromSnapshot(snap: { connectorHealth?: unknown }): number {
-  const healthRaw = snap.connectorHealth;
-  const health: ConnectorHealthRow[] = Array.isArray(healthRaw)
-    ? (healthRaw as ConnectorHealthRow[])
-    : [];
-  if (health.length === 0) {
-    console.log("[warn] Connectors: none registered.");
-    return 1;
-  }
-  console.log("Connector health:");
-  for (const h of health) {
-    const id = typeof h.connectorId === "string" ? h.connectorId : "?";
-    const st = typeof h.state === "string" ? h.state : "?";
-    const mark = healthStateMark(st);
-    console.log(`  ${mark} ${id}: ${st}`);
-  }
-  const sev = worstHealthSeverity(health);
-  return sev === "ok" ? 0 : 1;
-}
-
-async function doctorRunGatewayRpcs(client: IPCClient): Promise<number> {
-  const ping = await client.call<{ uptime?: number }>("gateway.ping", {});
-  const uptime = typeof ping.uptime === "number" && Number.isFinite(ping.uptime) ? ping.uptime : 0;
-  console.log(`[ok] Gateway: IPC OK (uptime ~${String(Math.round(uptime / 1000))}s).`);
-
-  const val = await client.call<{ ok: boolean; errors: string[]; warnings: string[] }>(
-    "config.validate",
-    {},
-  );
-  let exit = doctorPrintConfigValidation(val);
-
-  const snap = await client.call<{
-    index?: { totalItems?: unknown };
-    connectorHealth?: unknown;
-  }>("diag.snapshot", {});
-  exit = Math.max(exit, doctorPrintIndexFromSnapshot(snap));
-  exit = Math.max(exit, doctorPrintHealthFromSnapshot(snap));
-  return exit;
-}
-
-export async function runDoctor(_args: string[]): Promise<void> {
-  const paths = getCliPlatformPaths();
-  let exit = 0;
-  exit = Math.max(exit, doctorPrintBunCheck());
-
-  console.log(`Data dir: ${paths.dataDir}`);
-  console.log(`Gateway state file: ${gatewayStatePath(paths)}`);
-
-  exit = Math.max(exit, doctorPrintVaultCheck());
-
-  const voiceCfg = loadVoiceConfigFromDir(paths.configDir);
-  const voiceLines = doctorVoiceLines(voiceCfg, {
-    which: (n) => Bun.which(n),
-    platform: platform() as "win32" | "darwin" | "linux",
-  });
-  for (const line of voiceLines) {
-    console.log(line);
-    if (line.startsWith("[fail]")) exit = Math.max(exit, 2);
-    else if (line.startsWith("[warn]")) exit = Math.max(exit, 1);
-  }
-
-  const state = await readGatewayState(paths);
-  if (state !== undefined && isProcessAlive(state.pid)) {
-    const client = new IPCClient(state.socketPath);
-    try {
-      await client.connect();
-      exit = Math.max(exit, await doctorRunGatewayRpcs(client));
-    } catch (e) {
-      const msg = e instanceof Error ? e.message : String(e);
-      console.log(`[fail] Gateway: IPC failed — ${msg}`);
-      exit = Math.max(exit, 2);
-    } finally {
-      await client.disconnect().catch(() => {});
-    }
-  } else if (state === undefined) {
-    console.log("[fail] Gateway: not running (no gateway.json — start with: nimbus start).");
-    exit = Math.max(exit, 2);
-  } else {
-    console.log(
-      `[fail] Gateway: stale state (pid ${String(state.pid)} is not running) — try nimbus stop or remove the state file.`,
-    );
-    exit = Math.max(exit, 2);
-  }
-
-  process.exitCode = exit;
-}
-
-// ─── Voice doctor helpers ────────────────────────────────────────────────────
-
-export type DoctorVoiceConfig = {
-  enabled: boolean;
-  whisperPath: string;
-  piperPath: string;
-  piperModel: string;
-};
-
-export type DoctorEnv = {
-  which: (name: string) => string | null;
-  platform: "win32" | "darwin" | "linux";
-};
-
-function doctorPiperLines(cfg: DoctorVoiceConfig, env: DoctorEnv): string[] {
-  if (cfg.piperPath === "" && cfg.piperModel === "") return [];
-  const lines: string[] = [];
-  const hasAbsPath = cfg.piperPath.includes("/") || cfg.piperPath.includes("\\");
-  const piperBinOk = cfg.piperPath !== "" && (hasAbsPath || env.which(cfg.piperPath) !== null);
-  if (!piperBinOk) {
-    lines.push(`[warn] Voice: piper_path is set but the binary was not found: ${cfg.piperPath}`);
-  }
-  if (cfg.piperModel === "") {
-    lines.push(
-      "[warn] Voice: piper_path is set but piper_model is empty — Piper TTS will not run.",
-    );
-  }
-  return lines;
-}
-
-export function doctorVoiceLines(cfg: DoctorVoiceConfig, env: DoctorEnv): string[] {
-  if (!cfg.enabled) return [];
-  const lines: string[] = [];
-
-  const whisperOk =
-    cfg.whisperPath !== "" || env.which("whisper-cli") !== null || env.which("main") !== null;
-  lines.push(
-    whisperOk
-      ? "[ok] Voice: whisper-cli is available."
-      : "[warn] Voice: whisper-cli not found on PATH and voice.whisper_path is unset — STT will not work.",
-  );
-
-  const ffmpegOk = env.which("ffmpeg") !== null;
-  lines.push(
-    ffmpegOk
-      ? "[ok] Voice: ffmpeg is on PATH."
-      : "[warn] Voice: ffmpeg not found on PATH — wake word detection requires ffmpeg for audio capture.",
-  );
-
-  if (env.platform === "darwin") {
-    lines.push("[ok] Voice: macOS `say` is always available.");
-  } else if (env.platform === "win32") {
-    lines.push("[ok] Voice: Windows SAPI via PowerShell is always available.");
-  } else {
-    const espeakOk = env.which("espeak-ng") !== null;
-    const spdSayOk = env.which("spd-say") !== null;
-    if (espeakOk) {
-      lines.push("[ok] Voice: Linux TTS via espeak-ng.");
-    } else if (spdSayOk) {
-      lines.push("[ok] Voice: Linux TTS via spd-say (espeak-ng preferred).");
-    } else {
-      lines.push(
-        "[warn] Voice: neither espeak-ng nor spd-say found on PATH — install one to enable TTS on Linux.",
-      );
-    }
-  }
-
-  lines.push(...doctorPiperLines(cfg, env));
-  return lines;
-}
-
-function applyVoiceKey(out: Partial<DoctorVoiceConfig>, key: string, val: string): void {
-  if (key === "enabled") out.enabled = val === "true";
-  else if (key === "whisper_path") out.whisperPath = val;
-  else if (key === "piper_path") out.piperPath = val;
-  else if (key === "piper_model") out.piperModel = val;
-}
-
-function parseVoiceTomlLines(lines: string[]): Partial<DoctorVoiceConfig> {
-  let inVoice = false;
-  const out: Partial<DoctorVoiceConfig> = {};
-  for (const line of lines) {
-    const hashIdx = line.indexOf("#");
-    const trimmed = (hashIdx >= 0 ? line.slice(0, hashIdx) : line).trim();
-    if (trimmed === "") continue;
-    if (trimmed.startsWith("[") && trimmed.endsWith("]")) {
-      inVoice = trimmed === "[voice]";
-      continue;
-    }
-    if (!inVoice) continue;
-    const eq = trimmed.indexOf("=");
-    if (eq <= 0) continue;
-    const key = trimmed.slice(0, eq).trim();
-    const valRaw = trimmed.slice(eq + 1).trim();
-    const val = valRaw.startsWith('"') && valRaw.endsWith('"') ? valRaw.slice(1, -1) : valRaw;
-    applyVoiceKey(out, key, val);
-  }
-  return out;
-}
-
-function loadVoiceConfigFromDir(configDir: string): DoctorVoiceConfig {
-  const tomlPath = join(configDir, "nimbus.toml");
-  const defaults: DoctorVoiceConfig = {
-    enabled: false,
-    whisperPath: "",
-    piperPath: "",
-    piperModel: "",
-  };
-  if (!existsSync(tomlPath)) return defaults;
-  try {
-    const src = readFileSync(tomlPath, "utf8");
-    return { ...defaults, ...parseVoiceTomlLines(src.split(/\r?\n/)) };
-  } catch {
-    return defaults;
-  }
+export function runDoctor(args: string[]): Promise<void> {
+  return coreRunDoctor(args, productionDeps);
 }
