+++ sonar-project.properties
@@ -62,7 +62,7 @@ sonar.javascript.lcov.reportPaths=coverage/lcov.info,packages/ui/coverage/lcov.i
 # `scripts/coverage-floor/exclusions.ts`; the directional parity check
 # (`bun run audit:exclusion-parity`) enforces every Sonar pattern has a
 # local-registry counterpart.
-sonar.coverage.exclusions=**/index/*-v[0-9]*-sql.ts,**/perf/fixtures/synthetic-*-trace.ts,**/perf/surfaces/**,**/packages/mcp-connectors/*/src/server.ts,packages/gateway/src/index.ts,packages/gateway/src/platform/sandbox/linux.ts,packages/gateway/src/platform/sandbox/darwin.ts,packages/gateway/src/platform/sandbox/win32.ts,packages/gateway/src/platform/sandbox/orphan-reap.ts,packages/gateway/src/platform/sandbox/sandbox-runner.ts,packages/gateway/src/platform/sandbox/sandbox-wrapper.ts,packages/gateway/src/platform/sandbox/index.ts,packages/gateway/src/connectors/lazy-mesh/slot.ts,packages/gateway/src/ipc/server/options.ts,packages/gateway/src/ipc/agent-invoke.ts,packages/gateway/src/ipc/workflow-invoke.ts,packages/gateway/src/ipc/connector-rpc-handlers/context.ts,packages/sdk/src/testing/sandbox-probe.ts,packages/gateway/src-native/**,packages/gateway/src/connectors/index.ts,packages/gateway/src/embedding/embedding-runtime.ts,packages/gateway/src/index/ranked-item.ts,packages/gateway/src/vault/nimbus-vault.ts,packages/gateway/src/db/query-guard-worker.ts,packages/gateway/src/embedding/embedding-worker.ts,packages/gateway/src/vault/ffi-ptr.ts,packages/gateway/src/platform/index.ts,packages/gateway/src/vault/factory.ts,packages/cli/src/index.ts,packages/client/src/index.ts,packages/client/src/stream-events.ts,packages/sdk/src/ipc/index.ts,packages/gateway/src/embedding/load-feature-extraction-pipeline.ts,packages/cli/src/commands/start.ts,packages/cli/src/commands/tui.tsx,packages/cli/src/lib/gateway-process.ts,packages/cli/src/commands/repl.ts
+sonar.coverage.exclusions=**/index/*-v[0-9]*-sql.ts,**/perf/fixtures/synthetic-*-trace.ts,**/perf/surfaces/**,**/packages/mcp-connectors/*/src/server.ts,packages/gateway/src/index.ts,packages/gateway/src/platform/sandbox/linux.ts,packages/gateway/src/platform/sandbox/darwin.ts,packages/gateway/src/platform/sandbox/win32.ts,packages/gateway/src/platform/sandbox/orphan-reap.ts,packages/gateway/src/platform/sandbox/sandbox-runner.ts,packages/gateway/src/platform/sandbox/sandbox-wrapper.ts,packages/gateway/src/platform/sandbox/index.ts,packages/gateway/src/connectors/lazy-mesh/slot.ts,packages/gateway/src/ipc/server/options.ts,packages/gateway/src/ipc/agent-invoke.ts,packages/gateway/src/ipc/workflow-invoke.ts,packages/gateway/src/ipc/connector-rpc-handlers/context.ts,packages/sdk/src/testing/sandbox-probe.ts,packages/gateway/src-native/**,packages/gateway/src/connectors/index.ts,packages/gateway/src/embedding/embedding-runtime.ts,packages/gateway/src/index/ranked-item.ts,packages/gateway/src/vault/nimbus-vault.ts,packages/gateway/src/db/query-guard-worker.ts,packages/gateway/src/embedding/embedding-worker.ts,packages/gateway/src/vault/ffi-ptr.ts,packages/gateway/src/platform/index.ts,packages/gateway/src/vault/factory.ts,packages/cli/src/index.ts,packages/client/src/index.ts,packages/client/src/stream-events.ts,packages/sdk/src/ipc/index.ts,packages/gateway/src/embedding/load-feature-extraction-pipeline.ts,packages/cli/src/commands/start.ts,packages/cli/src/commands/tui.tsx,packages/cli/src/lib/gateway-process.ts,packages/cli/src/commands/repl.ts,packages/cli/src/commands/doctor.ts
 
 # Synthetic trace fixtures share an intentional shape (seeded LCG +
 # tier-scaled page generator); SQL migration templates are inherently
